Renamed normalizeTextWrapper.js to normalizeWrapper.js
